third_party/blink/web_tests/platform/generic/vibration/vibration-exceptions-expected.txt

Tests that bad input throws exceptions in the Vibration API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS navigator.vibrate() threw exception TypeError: Failed to execute 'vibrate' on 'Navigator': 1 argument required, but only 0 present..
PASS navigator.vibrate(1, 2) is true
PASS navigator.vibrate([1], [2]) is true
PASS successfullyParsed is true

TEST COMPLETE
